make some fns public
