+++ src/fmmax/_misc.py
@@ -8,11 +8,6 @@ from typing import Tuple
 import jax.numpy as jnp
 
 
-def angular_frequency_for_wavelength(wavelength: jnp.ndarray) -> jnp.ndarray:
-    """Returns the angular frequency for the specified wavelength."""
-    return 2 * jnp.pi / wavelength  # Since by our convention c == 1.
-
-
 def diag(x: jnp.ndarray) -> jnp.ndarray:
     """A batch-compatible version of `numpy.diag`."""
     shape = x.shape + (x.shape[-1],)
@@ -42,18 +37,3 @@ def atleast_nd(x: jnp.ndarray, n: int) -> jnp.ndarray:
     """Adds leading dimensions to `x`, ensuring that it is at least n-dimensional."""
     dims_to_add = tuple(range(max(0, n - x.ndim)))
     return jnp.expand_dims(x, axis=dims_to_add)
-
-
-def absolute_axes(axes: Tuple[int, ...], ndim: int) -> Tuple[int, ...]:
-    """Returns the absolute axes for given relative axes and array dimensionality."""
-    if not all(a in list(range(-ndim, ndim)) for a in axes):
-        raise ValueError(
-            f"All elements of `axes` must be in the range ({ndim}, {ndim - 1}) "
-            f"but got {axes}."
-        )
-    absolute_axes = tuple([d % ndim for d in axes])
-    if len(absolute_axes) != len(set(absolute_axes)):
-        raise ValueError(
-            f"Found duplicates in `axes`; computed absolute axes are {absolute_axes}."
-        )
-    return absolute_axes
