+++ src/fmmax/basis.py
@@ -12,7 +12,7 @@ import jax.numpy as jnp
 import numpy as onp
 from jax import tree_util
 
-from fmmax import _misc
+from fmmax import utils
 
 # Officially defines the x- and y- directions. By convention, the x-axis preceeds
 # the y-axis in our array indexing scheme.
@@ -208,7 +208,7 @@ def plane_wave_in_plane_wavevector(
     Returns:
         The fundamental transverse wavevector, i.e. ``(kx0, ky0)``.
     """
-    angular_frequency = _misc.angular_frequency_for_wavelength(wavelength)
+    angular_frequency = utils.angular_frequency_for_wavelength(wavelength)
     kx0 = (
         angular_frequency
         * jnp.sin(polar_angle)
